+++ src/App.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function App() {
+  return <h1>LUMIN</h1>;
+}
+
+export default App;
